move 4in2 partial driver to its own file
