+++ drivers/drivers_partial.py
@@ -595,556 +595,3 @@ class EPD2in13d(WavesharePartial):
             self.display_partial(self.get_frame_buffer(image), x, y, x + image.width, x + image.height)
         else:
             self.display_full(self.get_frame_buffer(image))
-
-
-class EPD4in2(WavesharePartial):
-    """Waveshare 4.2" """
-
-    # code adapted from  EPD_4in2.c
-
-    def __init__(self):
-        super().__init__(name='4.2"', width=300, height=400)
-
-    lut_vcom0 = [
-        0x00, 0x17, 0x00, 0x00, 0x00, 0x02,
-        0x00, 0x17, 0x17, 0x00, 0x00, 0x02,
-        0x00, 0x0A, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x0E, 0x0E, 0x00, 0x00, 0x02,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00,
-    ]
-
-    lut_ww = [
-        0x40, 0x17, 0x00, 0x00, 0x00, 0x02,
-        0x90, 0x17, 0x17, 0x00, 0x00, 0x02,
-        0x40, 0x0A, 0x01, 0x00, 0x00, 0x01,
-        0xA0, 0x0E, 0x0E, 0x00, 0x00, 0x02,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    lut_bw = [
-        0x40, 0x17, 0x00, 0x00, 0x00, 0x02,
-        0x90, 0x17, 0x17, 0x00, 0x00, 0x02,
-        0x40, 0x0A, 0x01, 0x00, 0x00, 0x01,
-        0xA0, 0x0E, 0x0E, 0x00, 0x00, 0x02,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    lut_wb = [
-        0x80, 0x17, 0x00, 0x00, 0x00, 0x02,
-        0x90, 0x17, 0x17, 0x00, 0x00, 0x02,
-        0x80, 0x0A, 0x01, 0x00, 0x00, 0x01,
-        0x50, 0x0E, 0x0E, 0x00, 0x00, 0x02,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    lut_bb = [
-        0x80, 0x17, 0x00, 0x00, 0x00, 0x02,
-        0x90, 0x17, 0x17, 0x00, 0x00, 0x02,
-        0x80, 0x0A, 0x01, 0x00, 0x00, 0x01,
-        0x50, 0x0E, 0x0E, 0x00, 0x00, 0x02,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    # luts for partial screen updates
-
-    partial_lut_vcom1 = [
-        0x00, 0x19, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00
-    ]
-
-    partial_lut_ww1 = [
-        0x00, 0x19, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    partial_lut_bw1 = [
-        0x80, 0x19, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    partial_lut_wb1 = [
-        0x40, 0x19, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    partial_lut_bb1 = [
-        0x00, 0x19, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    # gray
-
-    # 0~3 gray
-    gray_lut_vcom = [
-        0x00, 0x0A, 0x00, 0x00, 0x00, 0x01,
-        0x60, 0x14, 0x14, 0x00, 0x00, 0x01,
-        0x00, 0x14, 0x00, 0x00, 0x00, 0x01,
-        0x00, 0x13, 0x0A, 0x01, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-    ]
-    # R21
-    gray_lut_ww = [
-        0x40, 0x0A, 0x00, 0x00, 0x00, 0x01,
-        0x90, 0x14, 0x14, 0x00, 0x00, 0x01,
-        0x10, 0x14, 0x0A, 0x00, 0x00, 0x01,
-        0xA0, 0x13, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-    # R22H	r
-    gray_lut_bw = [
-        0x40, 0x0A, 0x00, 0x00, 0x00, 0x01,
-        0x90, 0x14, 0x14, 0x00, 0x00, 0x01,
-        0x00, 0x14, 0x0A, 0x00, 0x00, 0x01,
-        0x99, 0x0C, 0x01, 0x03, 0x04, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-    # R23H	w
-    gray_lut_wb = [
-        0x40, 0x0A, 0x00, 0x00, 0x00, 0x01,
-        0x90, 0x14, 0x14, 0x00, 0x00, 0x01,
-        0x00, 0x14, 0x0A, 0x00, 0x00, 0x01,
-        0x99, 0x0B, 0x04, 0x04, 0x01, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-    # R24H	b
-    gray_lut_bb = [
-        0x80, 0x0A, 0x00, 0x00, 0x00, 0x01,
-        0x90, 0x14, 0x14, 0x00, 0x00, 0x01,
-        0x20, 0x14, 0x0A, 0x00, 0x00, 0x01,
-        0x50, 0x13, 0x01, 0x00, 0x00, 0x01,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ]
-
-    ############### all of these override existing methods, they are
-    ############### implemented differently in the c file...
-
-    def reset(self):
-        self.digital_write(self.RST_PIN, 0x01)
-        self.delay_ms(200)
-        self.digital_write(self.RST_PIN, 0x00)
-        self.delay_ms(200)
-        self.digital_write(self.RST_PIN, 0x01)
-        self.delay_ms(200)
-
-    def send_command(self, command):
-        self.digital_write(self.DC_PIN, 0x00)
-        self.digital_write(self.CS_PIN, 0x00)
-        self.spi_transfer([command])
-        self.digital_write(self.CS_PIN, 0x01)
-
-    def send_data(self, data):
-        self.digital_write(self.DC_PIN, 0x01)
-        self.digital_write(self.CS_PIN, 0x00)
-        self.spi_transfer([data])
-        self.digital_write(self.CS_PIN, 0x01)
-
-    def wait_until_idle(self):
-        self.send_command(0x71)
-        while self.digital_read(self.BUSY_PIN) == 0:
-            self.send_command(0x71)
-            self.delay_ms(100)
-
-    def turn_on_display(self):
-        self.send_command(0x12)
-        self.delay_ms(100)
-        self.wait_until_idle()
-
-    def full_set_lut(self):
-        self.send_command(0x20)
-        for i in range(0, len(self.lut_vcom0)):
-            self.send_data(self.lut_vcom0[i])
-
-        self.send_command(0x21)
-        for i in range(0, len(self.lut_ww)):
-            self.send_data(self.lut_ww[i])
-
-        self.send_command(0x22)
-        for i in range(0, len(self.lut_bw)):
-            self.send_data(self.lut_bw[i])
-
-        self.send_command(0x23)
-        for i in range(0, len(self.lut_wb)):
-            self.send_data(self.lut_wb[i])
-
-        self.send_command(0x24)
-        for i in range(0, len(self.lut_bb)):
-            self.send_data(self.lut_bb[i])
-
-    def partial_set_lut(self):
-        self.send_command(0x20)
-        for i in range(0, len(self.partial_lut_vcom1)):
-            self.send_data(self.partial_lut_vcom1[i])
-
-        self.send_command(0x21)
-        for i in range(0, len(self.partial_lut_ww1)):
-            self.send_data(self.partial_lut_ww1[i])
-
-        self.send_command(0x22)
-        for i in range(0, len(self.partial_lut_bw1)):
-            self.send_data(self.partial_lut_bw1[i])
-
-        self.send_command(0x23)
-        for i in range(0, len(self.partial_lut_wb1)):
-            self.send_data(self.partial_lut_wb1[i])
-
-        self.send_command(0x24)
-        for i in range(0, len(self.partial_lut_bb1)):
-            self.send_data(self.partial_lut_bb1[i])
-
-    def gray_set_lut(self):
-        self.send_command(0x20)
-        for i in range(0, len(self.gray_lut_vcom)):
-            self.send_data(self.gray_lut_vcom[i])
-
-        self.send_command(0x21)
-        for i in range(0, len(self.gray_lut_ww)):
-            self.send_data(self.gray_lut_ww[i])
-
-        self.send_command(0x22)
-        for i in range(0, len(self.gray_lut_bw)):
-            self.send_data(self.gray_lut_bw[i])
-
-        self.send_command(0x23)
-        for i in range(0, len(self.gray_lut_wb)):
-            self.send_data(self.gray_lut_wb[i])
-
-        self.send_command(0x24)
-        for i in range(0, len(self.gray_lut_bb)):
-            self.send_data(self.gray_lut_bb[i])
-
-    def init_full(self):
-        self.reset()
-
-        self.send_command(0x01)  # POWER SETTING
-        self.send_data(0x03)
-        self.send_data(0x00)
-        self.send_data(0x2b)
-        self.send_data(0x2b)
-
-        self.send_command(0x06)  # boost soft start
-        self.send_data(0x17)  # A
-        self.send_data(0x17)  # B
-        self.send_data(0x17)  # C
-
-        self.send_command(0x04) # power on
-        self.wait_until_idle()
-
-        self.send_command(0x00)  # panel setting
-        self.send_data(0xbf)  #  KW-BF   KWR-AF	BWROTP 0f	BWOTP 1f
-        self.send_data(0x0d)
-
-        self.send_command(0x30)  # PLL setting
-        self.send_data(0x3c)  # 3a 100HZ   29 150Hz 39 200HZ	31 171HZ
-
-        self.send_command(0x61)  # resolution setting
-        self.send_data((self.width >> 8) & 0xff)
-        self.send_data(self.width & 0xff)
-        self.send_data((self.height >> 8) & 0xff)
-        self.send_data(self.height & 0xff)
-
-        self.send_command(0x82)  # vcom_DC setting
-        self.send_data(0x28)
-
-        self.send_command(0X50)			# VCOM AND DATA INTERVAL SETTING
-        self.send_data(0x97) # WBmode:VBDF 17|D7 VBDW 97 VBDB 57		WBRmode:VBDF F7 VBDW 77 VBDB 37  VBDR B7
-
-        self.full_set_lut()
-
-    def init_gray(self):
-        #### NOTE: This code is currently not being called.
-        self.reset()
-
-        self.send_command(0x01)  # POWER SETTING
-        self.send_data(0x03)
-        self.send_data(0x00)
-        self.send_data(0x2b)
-        self.send_data(0x2b)
-        self.send_data(0x13)
-
-        self.send_command(0x06)  # boost soft start
-        self.send_data(0x17)  # A
-        self.send_data(0x17)  # B
-        self.send_data(0x17)  # C
-
-        self.send_command(0x04)
-        self.wait_until_idle()
-
-        self.send_command(0x00)  # panel setting
-        self.send_data(0x3f)     # KW-3f   KWR-2F	BWROTP 0f	BWOTP 1f
-
-        self.send_command(0x30)  # PLL setting
-        self.send_data(0x3c)  # 3a 100HZ   29 150Hz 39 200HZ	31 171HZ
-
-        self.send_command(0x61)  # resolution setting
-        self.send_data((self.width >> 8) & 0xff)
-        self.send_data(self.width & 0xff)
-        self.send_data((self.height >> 8) & 0xff)
-        self.send_data(self.height & 0xff)
-
-        self.send_command(0x82)  # vcom_DC setting
-        self.send_data(0x12)
-
-        self.send_command(0X50)			#VCOM AND DATA INTERVAL SETTING
-        self.send_data(0x97)		#WBmode:VBDF 17|D7 VBDW 97 VBDB 57		WBRmode:VBDF F7 VBDW 77 VBDB 37  VBDR B7
-
-    def init(self, partial=True):
-        self.partial_refresh = partial
-        if self.epd_init() != 0:
-            return -1
-        self.init_full()
-
-    def clear(self):
-        width = int(self.width // 8 if self.width % 8 == 0 else self.width % 8 + 1)
-        height = int(self.height)
-
-        self.send_command(0x10)
-        for j in range(height):
-            for i in range(width):
-                self.send_data(0xff)
-
-        self.send_command(0x13)
-        for j in range(height):
-            for i in range(width):
-                self.send_data(0xff)
-
-        self.send_command(0x12) # Display refresh
-        self.delay_ms(10)
-        self.turn_on_display()
-
-    def display_full(self, frame_buffer):
-        if not frame_buffer:
-            return
-
-        width = int(self.width // 8 if self.width % 8 == 0 else self.width % 8 + 1)
-        height = int(self.height)
-
-        self.send_command(0x13)
-        for j in range(height):
-            for i in range(width):
-                self.send_data(frame_buffer[i + j * width])
-
-        self.turn_on_display()
-
-    def display_partial(self, frame_buffer, x_start, y_start, x_end, y_end):
-        if not frame_buffer:
-            return
-
-        width = int(self.width // 8 if self.width % 8 == 0 else self.width % 8 + 1)
-        height = int(self.height)
-
-        x_start = int(x_start if x_start % 8 == 0 else x_start // 8 * 8 + 8)
-        x_end = int(x_end if x_end % 8 == 0 else x_end // 8 * 8 + 8)
-
-        y_start = int(y_start)
-        y_end = int(y_end)
-
-        self.send_command(0x50)
-        self.send_data(0xf7)
-        self.delay_ms(100)
-
-        self.send_command(0x82) #vcom_DC setting
-        self.send_data(0x08)
-        self.send_command(0X50)
-        self.send_data(0x47)
-        self.partial_set_lut()
-        self.send_command(0x91)   # This command makes the display enter partial mode
-        self.send_command(0x90)   # resolution setting
-        self.send_data((x_start)//256)
-        self.send_data((x_start) % 256)   # x-start
-
-        self.send_data((x_end) // 256)
-        self.send_data((x_end) % 256 - 1)   # x-end
-
-        self.send_data(y_start // 256)
-        self.send_data(y_start % 256)   # y-start
-
-        self.send_data(y_end // 256)
-        self.send_data(y_end % 256 - 1)   # y-end
-        self.send_data(0x28)
-
-        self.send_command(0x10)   # writes Old data to SRAM for programming
-        for j in range(y_end - y_start):
-            for i in range((x_end - x_start) // 8):
-                self.send_data(frame_buffer[(y_start + j) * width + x_start // 8 + i])
-
-        self.send_command(0x13)   # writes New data to SRAM.
-        for j in range(y_end - y_start):
-            for i in range((x_end - x_start) // 8):
-                # there is an issue, because there are no unsigned values in
-                # python, not sure if it matters though
-                self.send_data(~frame_buffer[(y_start + j) * width + x_start // 8 + i])
-
-        self.send_command(0x12)   # DISPLAY REFRESH
-        self.delay_ms(10) # The delay here is necessary, 200uS at least!!!
-        self.turn_on_display()
-
-    def display_gray(self, frame_buffer):
-        #### NOTE: This code is currently not being called.
-        #### this is what the original source code says:
-        # /****Color display description****
-        #       white  gray1  gray2  black
-        # 0x10|  01     01     00     00
-        # 0x13|  01     00     01     00
-        # *********************************/
-        # 	EPD_4IN2_SendCommand(0x10);
-
-        self.send_command(0x10)
-
-        for m in range(self.height):
-            for i in range(self.width // 8):
-                temp3 = 0
-                for j in range(2):
-                    temp1 = frame_buffer[(m * (self.width // 8) + i) * 2 + j]
-                    for k in range(2):
-                        temp2 = temp1 & 0xC0
-                        if temp2 == 0xC0:
-                            temp3 |= 0x01 # white
-                        elif temp2 == 0x00:
-                            temp3 |= 0x00 # black
-                        elif temp2 == 0x80:
-                            temp3 |= 0x01 # gray1
-                        else: # 0x40
-                            temp3 |= 0x00 # gray2
-                        temp3 <<= 1
-
-                        temp1 <<= 2
-                        temp2 = temp1 & 0xC0
-                        if temp2 == 0xC0:  # white
-                            temp3 |= 0x01
-                        elif temp2 == 0x00: # black
-                            temp3 |= 0x00
-                        elif temp2 == 0x80:
-                            temp3 |= 0x01 # gray1
-                        else: # 0x40
-                            temp3 |= 0x00 # gray2
-                        if (j != 1) or (k != 1):
-                            temp3 <<= 1
-
-                        temp1 <<= 2
-                    # end for k
-                # end for j
-                self.send_data(temp3)
-            # end for i
-        # end for m
-        # new data
-        self.send_command(0x13)
-
-        for m in range(self.height):
-            for i in range(self.width // 8):
-                temp3 = 0
-                for j in range(2):
-                    temp1 = frame_buffer[(m * (self.width // 8) + i) * 2 + j]
-                    for k in range(2):
-                        temp2 = temp1 & 0xC0
-                        if temp2 == 0xC0:
-                            temp3 |= 0x01 # white
-                        elif temp2 == 0x00:
-                            temp3 |= 0x00 # black
-                        elif temp2 == 0x80:
-                            temp3 |= 0x00 # gray1
-                        else: # 0x40
-                            temp3 |= 0x01 # gray2
-                        temp3 <<= 1
-
-                        temp1 <<= 2
-                        temp2 = temp1 & 0xC0
-                        if temp2 == 0xC0:  # white
-                            temp3 |= 0x01
-                        elif temp2 == 0x00: # black
-                            temp3 |= 0x00
-                        elif temp2 == 0x80:
-                            temp3 |= 0x00 # gray1
-                        else:    # 0x40
-                            temp3 |= 0x01	# gray2
-                        if (j != 1) or (k != 1):
-                            temp3 <<= 1
-
-                        temp1 <<= 2
-                    # end for k
-                # end for j
-                self.send_data(temp3)
-            # end for i
-        # end for m
-        self.gray_set_lut()
-        self.turn_on_display()
-
-    def sleep(self):
-        self.send_command(0x02) # power off
-        self.wait_until_idle()
-        self.send_command(0x07) # deep sleep
-        self.send_data(0xa5)
-
-    # the implementation in super() is probably buggy!!!
-    def get_frame_buffer(self, image):
-        buf = [0x00] * int(self.width * self.height / 8)
-        # Set buffer to value of Python Imaging Library image.
-        # Image must be in mode 1.
-        image_monocolor = image.convert('1')
-        imwidth, imheight = image_monocolor.size
-        # if imwidth != self.width or imheight != self.height:
-        #     raise ValueError('Image must be same dimensions as display \
-        #         ({0}x{1}).'.format(self.width, self.height))
-
-        pixels = image_monocolor.load()
-        for y in range(imheight):
-            for x in range(imwidth):
-                # Set the bits for the column of pixels at the current position.
-                if pixels[x, y] != 0:
-                    buf[int((x + y * self.width) / 8)] |= 0x80 >> (x % 8)
-        return buf
-
-    def draw(self, x, y, image):
-        """Replace a particular area on the display with an image"""
-        if self.partial_refresh:
-            self.display_partial(self.get_frame_buffer(image),
-                                 x, y,
-                                 x + image.width, x + image.height)
-        else:
-            self.display_full(self.get_frame_buffer(image))
