+++ papertty.py
@@ -17,6 +17,7 @@ import drivers.drivers_full as drivers_full
 import drivers.drivers_color as drivers_color
 import drivers.drivers_colordraw as drivers_colordraw
 import drivers.driver_it8951 as driver_it8951
+import drivers.drivers_4in2 as driver_4in2
 
 # for ioctl
 import fcntl
@@ -400,7 +401,7 @@ def get_drivers():
     driverdict = {}
     driverlist = [drivers_partial.EPD1in54, drivers_partial.EPD2in13, drivers_partial.EPD2in13v2, drivers_partial.EPD2in9,
                   drivers_partial.EPD2in13d,
-                  drivers_full.EPD2in7, drivers_partial.EPD4in2, drivers_full.EPD7in5, drivers_full.EPD7in5v2,
+                  drivers_full.EPD2in7, driver_4in2.EPD4in2, drivers_full.EPD7in5, drivers_full.EPD7in5v2,
                   drivers_color.EPD4in2b, drivers_color.EPD7in5b, drivers_color.EPD5in83, drivers_color.EPD5in83b,
                   drivers_colordraw.EPD1in54b, drivers_colordraw.EPD1in54c, drivers_colordraw.EPD2in13b,
                   drivers_colordraw.EPD2in7b, drivers_colordraw.EPD2in9b, driver_it8951.IT8951,
